refactor(lib): extract lib/demo_seed.php (ADR-004, A14) — closes #909
